fix: rename challeng.tsx to challenge.tsx
