+++ apps/host/src/components/modules/challenges/challenge-grid/challenge.tsx
